WoT restructure D1b+D2: table.ts -> table/baseTable.ts + region subclasses -> table/regions/*Table.ts
