+++ handsontable/src/3rdparty/walkontable/src/core/_base.ts
@@ -1,6 +1,6 @@
 import type { DomBindings, WalkontableInstance } from '../types';
 import type Settings from '../settings';
-import type Table from '../table';
+import type Table from '../table/baseTable';
 import type Viewport from '../viewport/viewport';
 import type Overlays from '../overlays';
 import type { SelectionManager } from '../selection/manager';
