+++ handsontable/src/3rdparty/walkontable/src/core/core.ts
@@ -2,8 +2,8 @@ import Event, { createEventDeps } from '../event';
 import Overlays, { createOverlaysDeps } from '../overlays';
 import { CLONE_TYPES } from '../overlay';
 import Settings from '../settings';
-import MasterTable from '../table/master';
-import { createTableDeps } from '../table';
+import MasterTable from '../table/regions/masterTable';
+import { createTableDeps } from '../table/baseTable';
 import Viewport, { createViewportDeps } from '../viewport/viewport';
 import CoreAbstract from './_base';
 import { SelectionManager } from '../selection/manager';
