+++ handsontable/src/3rdparty/walkontable/src/overlay/topInlineStartCorner.ts
@@ -1,9 +1,9 @@
-import type { TableDeps } from '../table';
+import type { TableDeps } from '../table/baseTable';
 import {
   setOverlayPosition,
   resetCssTransform,
 } from '../../../../helpers/dom/element';
-import TopInlineStartCornerOverlayTable from '../table/topInlineStartCorner';
+import TopInlineStartCornerOverlayTable from '../table/regions/topInlineStartCornerTable';
 import { Overlay, type OverlayDeps } from './_base';
 import {
   CLONE_TOP_INLINE_START_CORNER,
