+++ handsontable/src/3rdparty/walkontable/src/table/regions/bottomTable.ts
@@ -1,9 +1,9 @@
-import type { TableDeps } from '../table';
-import Table from '../table';
-import stickyRowsBottom from '../rangeQuery/stickyRowsBottom';
-import { columnRangeQuery } from '../rangeQuery/virtualRange';
-import { mixin } from '../../../../helpers/object';
-import { CLONE_BOTTOM } from '../overlay';
+import type { TableDeps } from '../baseTable';
+import Table from '../baseTable';
+import stickyRowsBottom from '../../rangeQuery/stickyRowsBottom';
+import { columnRangeQuery } from '../../rangeQuery/virtualRange';
+import { mixin } from '../../../../../helpers/object';
+import { CLONE_BOTTOM } from '../../overlay';
 
 /**
  * Subclass of `Table` that provides the helper methods relevant to BottomOverlay, implemented through mixins.
