+++ handsontable/src/3rdparty/walkontable/src/table/regions/inlineStartTable.ts
@@ -1,9 +1,9 @@
-import type { TableDeps } from '../table';
-import Table from '../table';
-import { rowRangeQuery } from '../rangeQuery/virtualRange';
-import stickyColumnsStart from '../rangeQuery/stickyColumnsStart';
-import { mixin } from '../../../../helpers/object';
-import { CLONE_INLINE_START } from '../overlay';
+import type { TableDeps } from '../baseTable';
+import Table from '../baseTable';
+import { rowRangeQuery } from '../../rangeQuery/virtualRange';
+import stickyColumnsStart from '../../rangeQuery/stickyColumnsStart';
+import { mixin } from '../../../../../helpers/object';
+import { CLONE_INLINE_START } from '../../overlay';
 
 /**
  * Subclass of `Table` that provides the helper methods relevant to InlineStartOverlayTable, implemented through mixins.
