+++ handsontable/src/3rdparty/walkontable/src/table/regions/topTable.ts
@@ -1,9 +1,9 @@
-import type { TableDeps } from '../table';
-import Table from '../table';
-import stickyRowsTop from '../rangeQuery/stickyRowsTop';
-import { columnRangeQuery } from '../rangeQuery/virtualRange';
-import { mixin } from '../../../../helpers/object';
-import { CLONE_TOP } from '../overlay';
+import type { TableDeps } from '../baseTable';
+import Table from '../baseTable';
+import stickyRowsTop from '../../rangeQuery/stickyRowsTop';
+import { columnRangeQuery } from '../../rangeQuery/virtualRange';
+import { mixin } from '../../../../../helpers/object';
+import { CLONE_TOP } from '../../overlay';
 
 /**
  * Subclass of `Table` that provides the helper methods relevant to TopOverlay, implemented through mixins.
